feat(head): add encoder support
THe head is the first subsystem to get support for encoders. That means
that it brings up timer 2 and 3 to read its encoders (in
motor_hardware_common.c) and links in the motor control hardware
interface callbacks to handle encoder overflow, and hooks up responding
to all the relevant messages.