feat(motor-control,head,gripper): encoder bringup (#406)
* feat(motor-control): encoders in hardware

Add encoders to the hardware interfaces.

The encoders now present data as int32_t numbers of pulses. This number
is not just the currently-held value in the timer registers, but also
the sum of the net number of overflows, which is handled by adding an
encoder_overflow callback that must be called from the timer overflow
interrupt.

The encoder timer system works by using quadrature encoders to drive a
timer up or downcounter. That means that at any given time, barring
overflows, the value currently held in the timer counter register
reflects the current absolute position of the encoder, as a uint16_t.

When that position overflows high or low, an interrupt occurs and the
counter register wraps over in the specified direction. We can turn that
into effectively the high 16 bits of a 32 bit counter - an overflow
positive is like the number going from 2^16-1 to 2^16, so we add 1 to
the high bits; an overflow negative is like the number going from 0 to
-1, so we subtract a bit and let the signed math do the work.

Once we've formed the signed upper part of the 32 bit number, we can add
the unsigned lower part to it, and everything magically works.

This all has to happen in the hardware interface because it needs to be
accessible from multiple threads. This is, in fact, bad and not the
right thing to do; but since these are machine words it should more or
less work, and the fix is pretty involved - this number (and the step
position) should be exposed via mailboxes or notifications.

* feat(motor-control): handle encoder messages, data

Now that we have readable encoders (once individual axis controllers
actually integrate and bring up the hardware), we can integrate them
into our message handling. Ack messages have encoder pulse counts, which
get sent to can as positions in micrometers based on the linear motion
system config; the controllers can also answer queries from can about
the current status of the encoders.

* refactor(gantry): encoder config

These configurations, at least in simulators, let us test the configs.
We also need to add the relevant message handlers to everything involved
so stuff actually compiles. However, this hardware isn't hooked up to
anything, and the encoder position will always be 0.

* feat(head): add encoder support

THe head is the first subsystem to get support for encoders. That means
that it brings up timer 2 and 3 to read its encoders (in
motor_hardware_common.c) and links in the motor control hardware
interface callbacks to handle encoder overflow, and hooks up responding
to all the relevant messages.

* feat(gripper): bring up encoders

The gripper also has encoders that need to have their timers enabled and
their callbacks linked in. This is more of a holdover from other prs; a
followup pr that brings up the jaw encoders for actual use is certainly needed.

* feat(motor-control): add sq31_31*int32=int32

new fixed_point_multiply function is necessary for the signed integer
outputs of the encoder subsystem.

* feat(can): add EncoderPositionRequest/Response

These can be used to get the current output of the encoder without
needing to wait for a move to complete.

* refactor(can): encoder positions are signed

This is necessary because encoders can overrun a little when moving
negative to the stops. We could do some fancy handling for this... or
allow them to be signed.

* refactor(motor-control): add encoders to lms

These can be used to translate encoder pulse counts into positions.

* feat(motor-control): get/reset encoder pulses

These use the hal timer interface to get the pulse count from a running
timer and reset it (and its overflow flag).