fix(gantry): fix rev1 direction and handle current better (#411)
* fix(gantry): gantry-y-rev1 has active-hi dir

This was true on proto and is still true on rev1 - it wasn't an
electrical design error, the motor's in a different orientation.

While we're at it, make sure we ahve appropriate config files for the c
hardware configuration.

* Do not prioritize hold current

Or in fact pay attention to it. The run/hold current levels are
fixed (for now).