refactor(motor-control): Add position status
The position status will be sent from motor interrupt handlers to the
motor control task(s) where it can be retrieved or manipulated with
canbus commands. This happens in places other than the move status task
because it may need to feed back into the motors or need to interact
with the rest of the system.