refactor(motor-control): better templatign for isr
The new template verification for the motor interrupt handler ISR is
needed to handle adding a concept for the move status reporter client,
since we also want to have a concept for the position status.

That means that the concept needs to know about which kind of message
this interrupt handler is using, which means it needs to be here, so a
bunch of stuff shifts around.