refactor(simulators): add command line options
We had a bunch of options we could pass to the simulators but they were
documented in source and docs and not in an inspectable way. It also
makes it annoying to add more.

Instead, or really in addition, we can use boost::program_options to
parse the environment variables. We can also really easily use it to
provide command line flags alongside them, and then we get a -h option
that can document both.

While we're at it, add a new option for the eeprom that specifies a file
to back up the eeprom contents.