refactor(eeprom): use a file to back up the eeprom
This lets eeprom changes be persistent between runs, if needed. We can
take advantage of the new options to specify a filename, and use normal
file handling utilities (not fstreams because they are slightly
nightmarish to deal with) to handle and write to files.

Also, by default the pipette simulator now uses a file backing for its
eeprom when you run with ./run_simulators.