refactor(sensors): smaller data in cap sensor sim
The simulator harness for the capacitive sensor returned numbers large
enough that the capdac would be adjusted, but then the harness didn't
take the altered capdac into account. That means the numbers returned on
can would change over time. By reducing the values returned from the
sensor, we fix the problem.