fix(pipettes): rev1: bring up M24128 eeprom (#441)
* fix(pipettes): rev1 pipettes use M24128 eeprom

Not the older kind. Explicitly set them to do so.

Closes RLIQ-97

* fix(i2c): retry transactions after NAK

An I2C NAK indicates that a device won't respond to a message. This can
be for basically any reason, including that it doesn't recognize it, but
one reason is that the device isn't ready yet. For instance, the M24128
eeprom will NAK all incoming messages during its write delay. That means
that the NAK can be a good way to get low-latency estimations of when a
device is ready to transmit.

By altering the i2c hardware interface code so that it retries a
transaction if it failed because it was NAK'd, we can use that
retryability to handle the write delay without an explicit sleep (though
we're using a lot of processor effort to do so.)

Closes RLIQ-98

* fix(i2c): allow messages big enough for M24128

This isn't really the right ultimate fix - we should make this allow
much larger messages, up to 64 bytes ideally, so we can write more in
one go - but it'll do for now.

Closes RLIQ-99