this needs some testing but the top line seems to fix the issue of homeing returning the wrong value sometimes, I think the root of that issue is noise on the limit switch, which starts the process of sending a move_complete, but the motor is still moving so by time the message is sent the encoder position is in a new place
doesn't entirely make sense becuase the next time the jaw closes it has the correct position, but i'm not ruleing out some
race condition for this.

the symptom is that during a gripper jaw home, (G axis) the position reported is not 0, instead being something like 10.

I think its also an option to get rid of the limit switch all together and use the encoder_g_speed detection like we do
for the grip command to detect the home position, and we can get rid of the limit switch all together for that axis