feat(pipettes): add specific can messages for 96 channel tip motors
Rather than having a specific node id for the tip motors on the 96 channel, I decided to create
specific CAN messages for certain motor functionalities. Unlike the gripper, the tip motors are
pretty intrinsic to the rest of the pipetting functionality and these CAN messages are only used in
very specific instances.