refactor(head): drive tool detect digitally
Tool detect is now done on pipettes only, based on logic level values.
The head sets those logic lvels on its gpios. This commit adds the logic
level setting, such that the left/z mount is high and the right/a mount
is low. The gripper/g mount is now also low, but it won't be used so
this just gets it out of the way.