refactor(pipettes,bootloader): digital mount id
We will no longer use the ADC to figure out a voltage to determine what
mount we're on; instead, we'll use a GPIO digital measurement. We can
still use most of the structure we wrote for this originally, but the
actual mechanism to decide the outcome is much much simpler.