refactor: use internal pull and inputs on head
Mount detection requires the head to be able to detect, which means its
pins need to be inputs. They can tell the pipette which one they are by
using internal pulls. This only works on DVT and up because EVT and down
has external pullups that are stronger than the internal ones, so on EVT
and down these pins need to actively drive the lines.