refactor: drive mount id for presence
Pipettes first detect which mount they're on (detecting the level
established by the head) and then drive the pin to the opposite level by
switching their pin mode to output. This will let the head know that
something is attached.

This only works on DVT and up because EVT and down require the head to
also be in output mode and don't have enough series resistance to handle
the head and the pipette driving in opposite directions.