refactor: add DVT-compliant presence sensing
The new presence sensing strategy is:
- The head sets its mount pins to input with internal pulls. The Z and
gripper mounts are pulled low, A is pulled high. This will cause the ID
pins on the mounts to be the value set by the pulls as long as no
pipette is connected or the connected pipette keeps its pins as inputs.
- When a pipette boots, it sets its mount id pins to input and reads the
value. The value it sees - which is determined by the internal pulls on
the head - tells the pipette which mount it's on.
- After the pipette finds its mount, it switches its mount id pin to
output and drives the line toward the value not pulled. For instance, if
the pipette determines it's on Z (because the pin is pulled low), it
drives the pin high. If the pipette determines it's on A (because the
pin is pulled high), it drives the pin low.
- When the head sees a mount pin change its value away from what's
pulled, it knows a pipette must be present and can send notifications.

There are some added complexities:
- It will likely take a little bit for the mount pin to regain its
pulled value after a pipette switches to input (e.g. when rebooting) or
when disconnecting; this means pipettes should wait after booting to
check the mount id value and drive the output to give time to settle
- The pipettes will have this sense/drive behavior both in the
bootloader and the application to allow for being connected in either
state, so when a pipette switches to bootloader the head may see it
spuriously disconnect and reconnect. python code should allow for this.