fix(pipettes): add correct microstepping for the tip motors
When we changed the motor drivers for the tip motors to use tmc2160s, we did not change the
microstepping back to 32. The issue hadn't been found thus far because the tip motors were also
executing moves twice so it would appear to be moving mostly to the correct position.