refactor: handle the DVT mount/pipette ID scheme (#551)
In EVT and previous, we had a scheme (or set of frequently-changing schemes) to have a passives network establish an appropriate voltage for a given (pipette type, attach point) pair that could be read by both the attached pipette and the head board. This kept having problems with rise time, consistency between assembled units, the quality of the onboard ADCs, and so on.

Instead, we'll eventually have a system where the head sets its mount id pins to input with internal pulls. This establishes a voltage the pipettes can see; they'll read it, and then they'll drive the pin to the opposite voltage, which the head can detect to indicate pipette presence.

However, this scheme won't work with EVT pipettes or head boards because of the presence of external pulls which are stronger than the internal pulls. That means the head board has to drive its detection pins directly, and that means the pipettes have to keep their mount ID pins on input so we're not driving current between two GPIOs. To get the above scheme to work, either both sides need to be DVTs or they need to be modified to remove the external pulls.

For now, and what's in this PR, what we're going to do is drive the pins from the head and keep the pipette pins inputs. This removes the ability for the head to detect the presence of a pipette, so we're going to keep this change in a branch sent to the factory to do QC of DVT pipettes and not distribute it further. We'll add, in a separate branch, the new behavior that allows detection again; and eventually, when we get more DVT hardware we'll either move away from the remaining EVT hardware or modify it to work with the new scheme.

What we have now is:

    The head drives its left/z mount id high
    The head drives its right/a mount id low
    The pipettes sense this and assign their ids from it
    Note: The above is switched from the first push of this PR because driving the voltage high on the left/z mount of EVT systems results in an unusable intermediate voltage because of the values of the resistor divider for the old analog system.
    Test cases:
    Pipette plugged into left at system boot
    Pipette plugged into right at system boot
    Pipette plugged into left after system boot
    Pipette plugged into right after system boot
    Pipette sent to bootloader on left
    Pipette sent to bootloader on right

Closes RET-1297