chore(cmake): add a new directory for cmake
We have the shared repository cmake-utils for cmake scripts that are
useful across projects, but sometimes we have cmake stuff that has a lot
of utility and is extremely specific to just ot3-firmware, but used in
several parts of ot3-firmware. That will be in common/cmake,  and
therefore we need that directory in our cmake module path. For now, it's
only useful in firmware.