feat(cmake): Macro for multi-revision builds
We need to build most of our firmware projects across several electrical
revisions. When we build those projects for each revision, almost
everything is the same, but there are a couple little differences we
need to handle - for instance, we need to be on a couple different
architectures, and some different revisions have different source
profiles.

This commit therefore introduces a couple new macros that will let us
avoid copy/pasting a whole bunch of code, and remove code that was
previously heavily copy/pasted.

ClangTidy is a pretty quick one; we had the same code that invokes
clangtidy, which is annoyingly specific, copy/pasted everywhere. The
macro in this module will let us avoid this.

The big one is the RevisionHandling module. This module defines a single
very large macro that lets you compile code and provide the same useful
targets for a project across multiple revisions. It consolidates the
copy/pasted code that provides things like
- .hex output!
- bootloader fold-in!
- flashing!
- debugging!
- C++ and C warning configurations!
- our basic include directory!
into one macro.

There are of course some things that need to avoid those defaults, and
therefore you can turn little bits off. There are also things that need
to be different between projects and can't be generalized, so you can
provide effectively a callback - the name of a macro that will be
executed once for each revision.

There's also the ability to add an alias for an old revision name (like,
for instance, -proto and -rev1 for -a1 and -b1) to ease the transition.