build: add function to add revision data to target
Previously we were adding revision data inside the foreach_revision
macro, but sometimes we're going to need revision data outside of that -
for instance, in tests. To avoid copying code, factor the revision
handling into its own function that "returns values" (sets them with a
prefix in parent scope) and use it in revision handling. Next, we'll use
it in other places.