refactor: add revision to device info
The revision of the PCBA that is currently running the firmware is part
of the firmware image. This seems circular, but it is established by the
firmware image that is programmed at the factory when the PCBA is
integrated into its subassembly - at that time, we'll program using an
image that has a bootloader with the appropriate revision compiled in.
From there, we'll only update micros with application firmware for the
revision that matches the bootloader. That means we don't need to store
the firmware revision in the eeprom, and compiling it in is enough.

The revision needs to be communicated from a microcontroller to the host
for this to work. That happens by adding it to the device info
message (with tests) for all the bootloaders and all the application
firmwares (all the ones that currently handle device info, anyway - it's
not implemented yet in rear-panel).

There will still be some bootloaders in the wild that don't and will
never have this change because it's impractical to reflash the boards,
but they are all DVT boards of various varieties and can be defaulted as
such in the python side of the update process.

Closes RET-1321