refactor: add revision to device info (#567)
* build: add function to add revision data to target

Previously we were adding revision data inside the foreach_revision
macro, but sometimes we're going to need revision data outside of that -
for instance, in tests. To avoid copying code, factor the revision
handling into its own function that "returns values" (sets them with a
prefix in parent scope) and use it in revision handling. Next, we'll use
it in other places.

* build: fix up c++-only compile options

Some compile options we use need to be present only if we're compiling a
c++ file since they're invalid for C. There were a bunch of places where
they weren't so limited, but were fine because those targets only
compiled C++. In a lot of places we'll now be adding the revision
internment c file (this is a lot nicer than having yet another setup
where we make a billion copies of a library) so make sure all these
options are properly scoped to the language.

* build: add revision info to tests and simulators

The simulators unambiguously need revision information (or will soon) -
we'll default them to a1, using the function we added previously. Most
tests will now also need revision information because they'll include
revision.h somewhere, so add things there too at the arbitrary revision
of a1.

* refactor: add revision to device info

The revision of the PCBA that is currently running the firmware is part
of the firmware image. This seems circular, but it is established by the
firmware image that is programmed at the factory when the PCBA is
integrated into its subassembly - at that time, we'll program using an
image that has a bootloader with the appropriate revision compiled in.
From there, we'll only update micros with application firmware for the
revision that matches the bootloader. That means we don't need to store
the firmware revision in the eeprom, and compiling it in is enough.

The revision needs to be communicated from a microcontroller to the host
for this to work. That happens by adding it to the device info
message (with tests) for all the bootloaders and all the application
firmwares (all the ones that currently handle device info, anyway - it's
not implemented yet in rear-panel).

There will still be some bootloaders in the wild that don't and will
never have this change because it's impractical to reflash the boards,
but they are all DVT boards of various varieties and can be defaulted as
such in the python side of the update process.

Closes RET-1321