refactor: deliver all firmware revisions to OE (#568)
* refactor(build): install subsystem revisions

Install all the revisions in the manifest.

* refactor(build): add no-dir-required presets

While the presets we currently have are really convenient for
developers, sometimes you really want to specify your own binary
directory and install directories, for instance during openembedded
builds. Preset version 3 supports this, so bump to require cmake 3.21
and move presets to version 3, so we can make a new
cross-no-directory-reqs preset that lets you _not_ specify the binary
dir in the preset, so OE can do it itself. This also lets us use a
slightly nicer method of setting install prefixes and toolchain files.

Closes RET-1322