fix(head): flip encoder on C2 (#582)
When we removed a gear in DVT, it flipped the direction of the lead screw relative to the motor shaft. We therefore swapped which direction we consider forward on the motor; but we didn't do it for the encoder, which is also on the motor shaft and therefore needs the same treatment. This PR swaps the encoder direction by swapping the active level of one of its pins.

Closes RET-1338