motor-control: errors for moves during estop
The motor controller would send an error if it was executing a move and
an estop happened; but if it estop was asserted and it got new moves, it
just wouldn't ever execute them. That means that anything that was
waiting on those moves would have to wait for some timeout, which in the
problematic case of the robot server starting while estop is asserted
would be forever for home moves.

This commit extends the behavior of "nak the first move and blackhole
the rest" to moves that come in while estop is asserted, which should
raise an error in the host side. The host side will have to be fixed to
handle such errors and still start up instead of exploding.

Closes RET-1340