refactor(include): modify the cap sensor task to take readings from different CINs
The eight channel pipette is unique insofar as there is only a single cap sensor with two flex
cables connected to two different channels. Since we had to change up some of the register values, I
decided to go ahead and refactor the code to be more uniform with the pressure/environment sensor.
During the refactor I realized that the baseline sensor request does the same thing as auto zero
threshold...(duh).. and we should refactor the code to 1. remove the baseline sensor request message
and 2. ensure that all of the sensors have the same functionality in terms of auto-zeroing.