fix(pipettes): ensure that secondary hardware exists before trying to access the value
I previously thought calling `.value()` on an optional parameter meant that it would either return a
nullptr or the object it was supposed to store. Mais non, I was wrong and therefore we have to do
this gross check to make sure we don't brick pipette firmware. If anyone has any elegant solutions,
definitely let me know.