fix(bootloader): present pipette information
The bootlaoder has no way to indicate what kind of pipette it is. This
means that if a pipette is stuck in its bootloader, it cannot be
automatically and safely updated.

To fix this, let's add a new byte to the device info and put the pipette
type in it (for a pipette, anyway). The byte will be interpreted
differently depending on the canbus node id that the device info
response comes from, so there's no need for a ton of space here.

This is implemented in the bootloader with another attribute-weak hook
that will only get compiled for the pipette itself. While I was here, I
refactored the way we do node IDs to make things a little better.

Closes RQA-584