fix(bootloader): delay before app start on pipette
On the pipettes, we don't reliably detect that we're on the right mount
when we go from bootloader to application. This is because it's too
fast. So let's slow it down. By reconfiguring the mount detect pin to
tristate and adding a delay after that before going to the app, we make
it reliable again.