perf(pipettes): remove interrupt initialization for the data ready line on the 96 channel
Due to the noise on the data ready lines, the 96 channel firmware gets overwhelmed with the addition
of the secondary sensor and crashes. For now, we should completely disable the initialization of the
data ready interrupt. The real fix is to use a software timer rather use EXTI on the data ready
line.