fix(motor-control): stop isr moves on StopRequest
When we got a StopRequest, we previously would clear the incoming queue
and then disable the motors, and didn't touch the interrupt handler -
because we only used StopRequests during estop, and that will hit the
interrupt handler just fine.

However, we now want to send stop requests from the host, and in that
case we need to stop a move that's possibly ongoing inside the interrupt
handler because the host will send followup messages that expect there
to be no move running.

We can use an atomic bool to implement a little event that will act the
same as the estop pinging.

Closes RET-1353