feat(include): add error code for overpressure trigger
We should throw an error message for the use-case of an overpressure spike being detected.
Potentially we'll want to do some filtering at some point, but I think the chances of us getting
into this value range is very low and thus it's OK to not do any type of signal processing.