feat(include): add checks for exceeding max sensor readings for the pressure and cap sensor (#672)
This is only really relevant to the pressure sensor, but I decided to add support for the cap sensor
as well. We want to make sure that the sensor does not ever exceed its max output to ensure that the
sensor itself doesn't break.


* feat(include): add error code for overpressure trigger (#673)

We should throw an error message for the use-case of an overpressure spike being detected.
Potentially we'll want to do some filtering at some point, but I think the chances of us getting
into this value range is very low and thus it's OK to not do any type of signal processing.